Adding df3 iterpolations 6-10 & query for df3 size.
Adding the last of the df3 interpolations in 6 through 10 I had in mind
in starting this update. These interpolations work alone, but are really
intended to be helper interpolations used with modes 0-5, which track
with the df3 file itself. Also added a capability for mode 6 to return
a df3 size by passing very specific, distant coordinates to the
density_file pattern. Looked into one of the TODOs in the continuous
pattern wave code and moved to EPSILON over previous 1e-5 magic value.