Adding soft_object pattern.
Like object pattern excepting an on the fly virtual grid is
used along with exponential blobbing of grid points within
the object to creat a pattern with a gradual 0 to 1 value
change.

Example use in pigment:

pigment {
    soft_object { Union00 } spacing 0.05 strength 0.95
    color_map { ColorMap6 }
}

Example use in a function for an isosurface etc:

function {
    pattern { soft_object { Text00 } spacing 0.009 strength 1.0 }
}

Parameters:

spacing - Optional. Virtual grid point to point space Defaults to 0.25.
          Clamped to a minimum of 1e-4.
strength - Optional. The blobbing strength. Defaults to 1.0. Normally use
           values of >0.0 to 1.0. Larger the 1.0 will be distorted, but as
           a pattern perhaps the distorted result is of use.