Adding hard_object pattern. Minor scrub of soft_object code.
Like object pattern excepting an on the fly value from the surface
inward is calculated with a value of 0.0 at the surface to 1.0 when
>= specified radius inside.

Example use in pigment:

pigment {
    hard_object { Union00 } radius 0.04 recursion_limit 10 samples 22
    color_map { ColorMap6 }
}

Example use in a function for use in an isosurface:

function {
    pattern { hard_object { Text00 }
        radius 0.04 recursion_limit 10 samples 22
    }
}

Parameters:

radius - >0.0. Depth inward from objust surface for 0 to 1 value ramp.
recursion_limit - 4+. Higher values better accuracy.
samples - 3 to 200. Number of spherical samples take in each recursion.