Update sbisect to return no roots when ray domain range is tiny.
Roots being very closely packed as happens most often when a ray is nearly
tangent to a surface cause instability in the sturm change sign counting
(numchanges) results. The instability, if not avoided, leads to incomplete
and sometimes inaccurately ordered roots. One of the triggers for the CEY
patch added in 1997 and where we get the diagonal lines of no roots in
certain objects.