Adding additional black_hole types 2 thru 7.
In addition to type 1 which fixes / clamps the inward black_hole, added
types 2,3 and 4 which are like type 1 except limiting results to x, y, and z
coordinate movements, respectively.

New types 5,6 and 7 represent new whirlpool warps are the localized
spherical coordinates x, y and z axis. The inverse keyword reverses the
normal direction of rotation. Setting strength 1.0 means a rotation of 2*pi
or a full rotation at the origin - which is a lot. Strengths of 0.5 perhaps
a better starting half rotation at the origin.